% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_down_i.R
\name{break_down}
\alias{break_down}
\title{Model Agnostic Sequential Variable Attributions}
\usage{
break_down(..., interactions = FALSE)
}
\arguments{
\item{...}{parameters passed to `local_*` functions}

\item{interactions}{shall interactions be included}
}
\value{
an object of the `break_down` class
}
\description{
This function finds Variable Attributions via Sequential Variable Conditioning
It calls either `local_attributions` or `local_interactions`.
}
\examples{
\dontrun{
## Not run:
library("DALEX")
library("iBreakDown")
library("randomForest")
set.seed(1313)
# example with interaction
# classification for HR data
model <- randomForest(status ~ . , data = HR)
new_observation <- HR_test[1,]

explainer_rf <- explain(model,
                        data = HR[1:1000,1:5],
                        y = HR$status[1:1000])

bd_rf <- break_down(explainer_rf,
                           new_observation)
bd_rf
plot(bd_rf, start_baseline = TRUE)
}
}
