% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_explainer.R
\name{convert_explainer}
\alias{convert_explainer}
\title{Converts a `break_down_uncertainty` to `break_down` explainer}
\usage{
convert_explainer(explainer)
}
\arguments{
\item{explainer}{explainer}
}
\description{
Called by `r describe.break_down_uncertainty`.
}
