% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_breakdown.R, R/describe_shap.R
\name{describe}
\alias{describe}
\alias{describe.break_down}
\alias{describe.break_down_uncertainty}
\title{Natural language description of a DALEX explainer}
\usage{
describe(explainer, nonsignificance_treshold = 0,15, ...)

\method{describe}{break_down}(explainer, nonsignificance_treshold = 0,15,
  ..., label = NULL, short_description = FALSE,
  display_values = FALSE, display_numbers = FALSE,
  display_distribution_details = FALSE, display_argumentation = 1,
  display_shap = FALSE)

\method{describe}{break_down_uncertainty}(explainer,
  nonsignificance_treshold = 0,15, ..., label = NULL,
  short_description = FALSE, display_values = FALSE,
  display_numbers = FALSE, display_distribution_details = FALSE,
  display_argumentation = 1, display_shap = FALSE)
}
\arguments{
\item{explainer}{a DALEX explainer}

\item{nonsignificance_treshold}{a parameter specifying a treshold for variable importance}

\item{...}{other arguments}

\item{label}{a short description describing model's prediction}

\item{short_description}{allows for generating short description}

\item{display_values}{allows for displaying variable values}

\item{display_numbers}{allows for displaying numerical values}

\item{display_distribution_details}{displays detailed description of predictions distribution}

\item{display_argumentation}{choosing the type of argumentation to be displayed}

\item{display_shap}{displays additional information about average contribution. Use only if the explainer is a shap explainer}
}
\value{
A string of natural language description of an explainer
}
\description{
Generic function \code{describe} generates a natural language description of
\code{break_down()} and \code{shap()}explanations what enchaces their interpretability.
}
\details{
Function \code{describe.break_down()} first selects one of four different description
scenarios. Depending on \code{nonsignificance_treshold} which selects variables with
significant contribution to model's final prediction, the scenario can be one of the following:
1. Model's prediction for the selected instance is significantly higher than the average prediction.
2. Model's prediction is significantly lower.
3. Model's prediction is close to it's average prediction, however there are significant
variables counteracting with each other
4. Model's prediction is close to it's average prediction and all the variables are rather nonsignificant.
For each scenario, a template is choosen. Currently only the defalut \code{argumentation_mode} can be choosen, which
chooses three most significant variables and dicribes them.

Arguments \code{display_values}, which shows values of the variables, and
\code{display_numbers}, which allows for displaying numerical values of prediciton changes, controls
the complexity of the generated description. Additional parameter \code{short_description} allows for generating short description,
while \code{display_distribution_details} displays additional details of the distribution of
model's prediction for all instances, if describing a break_down explanation. Currently only the default
mode of argumentation is supported.

Function \code{describe.break_down_uncertainty()} generates a natural language description
for a shap explanation. It imitates the \code{describe.break_down()}, the only difference is that contributions
are average contributions of iterated break_down contributions. Argument `display_shap` allows for
average's dispersion description.
}
\examples{
library("DALEX")
library("randomForest")
library("iBreakDown")

titanic <- na.omit(titanic)
model_titanic_rf <- randomForest(survived == "yes" ~ gender + age + class + embarked +
                                  fare + sibsp + parch,  data = titanic)

explain_titanic_rf <- explain(model_titanic_rf,
                              data = titanic[,-9],
                              y = titanic$survived == "yes",
                              label = "Random Forest v7")
set.seed(1234)
random_passanger <- titanic[sample(nrow(titanic),1),c(1,2,3,4,6,7,8)]
rf_la <- break_down(explain_titanic_rf, random_passanger, keep_distributions = TRUE)

description <- describe(rf_la,
                        label = "the passanger will survive with probability",
                        short_description = FALSE,
                        display_values =  TRUE,
                        display_numbers = TRUE,
                        display_distribution_details = FALSE,
                        display_argumentation = 1)

description




library("DALEX")
library("iBreakDown")
titanic <- na.omit(titanic)
model_titanic_glm <- glm(titanic$survived == "yes" ~ age + gender + class + fare + sibsp,
                         data = titanic[ ,-9], family = "binomial")
explain_titanic_glm <- explain(model_titanic_glm,
                              data = titanic[,-9],
                              y = titanic$survived == "yes",
                              label = "glm")
passanger <- titanic[sample(nrow(titanic), 1) ,-9]
shap_glm <- shap(explain_titanic_glm, passanger)
plot(shap_glm)

describe(shap_glm,
         label = "the selected passanger survives with probability",
         display_shap = TRUE,
         display_numbers = TRUE)


}
