% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_distributiondescription.R
\name{describe_distribution}
\alias{describe_distribution}
\title{Describes distribution of a model pedictions.}
\usage{
describe_distribution(explainer, display_numbers,
  display_distribution_details, model_name)
}
\arguments{
\item{explainer}{iBreakDown explainer}

\item{display_numbers}{TRUE for displaying variable contributions and intercept}

\item{display_distribution_details}{TRUE for displaying detailed description of predictions distribution}

\item{model_name}{name of the model to be explained}
}
\value{
a decription of predictions distribution
}
\description{
Describes distribution of a model pedictions.
}
