% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description_profile.R
\name{description_profile}
\alias{description_profile}
\title{Returns a scenario for a description}
\usage{
description_profile(explainer, nonsignificance_treshold = 0.15)
}
\arguments{
\item{explainer}{an iBreakDown explainer}

\item{nonsignificance_treshold}{a treshold for specyfiying which predictions are close to the
average model prediction. A prediction is significantly higher/lower
than the average prediction if
(average prediction - prediction)/(average prediction)
is higher than nonsignificance_treshold. It also has effect
on when a variables impact is considered as unimportant.}
}
\value{
an integer from 0 to 3
}
\description{
0 = prediction significantly higher than the average
1 = prediction significantly lower than the average
2 = prediction close to the average, but there are significant contributors
3 = prediction close to the average and there are no significant contributors
}
