% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_argument.R
\name{make_argument}
\alias{make_argument}
\title{Makes an argument for description}
\usage{
make_argument(explainer, label, display_values, display_numbers,
  display_argumentation, model_name, description_profile, display_shap)
}
\arguments{
\item{explainer}{an iBreakDown explanation}

\item{label}{a prediction label}

\item{display_values}{displays values}

\item{display_numbers}{displays numbers}

\item{display_argumentation}{displays argumentation mode}

\item{model_name}{name of the model to be explained}

\item{description_profile}{description scenario}

\item{display_shap}{displays addition information about shap explanation}
}
\value{
an argument
}
\description{
Makes an argument for description
}
