% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_introduction.R
\name{make_introduction}
\alias{make_introduction}
\title{Makes an introduction for the description function}
\usage{
make_introduction(explainer, label, display_numbers, distribution_details,
  model_name, distribution_kept, description_profile)
}
\arguments{
\item{explainer}{an iBreakDown explainer}

\item{label}{prediction description}

\item{display_numbers}{TRUE for displaying variable contributions and intercept}

\item{distribution_details}{TRUE for displaying detailed description of predictions distribution}

\item{model_name}{name of the model to be explained}

\item{distribution_kept}{TRUE if the distribution details are stored}

\item{description_profile}{description scenario}
}
\description{
Makes an introduction for the description function
}
