% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_short_description.R
\name{make_short_description}
\alias{make_short_description}
\title{Makes a short description of iBreakDown explainer}
\usage{
make_short_description(explainer, display_values, label, model_name,
  description_profile)
}
\arguments{
\item{explainer}{iBreakDown explainer}

\item{display_values}{if TRUE then displays variable valeus}

\item{label}{prediction label}

\item{model_name}{name of the model to be explained}

\item{description_profile}{a scenario for description}
}
\description{
Makes a short description of iBreakDown explainer
}
