% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_summary.R
\name{make_summary}
\alias{make_summary}
\title{Makes a summary for the description}
\usage{
make_summary(explainer, display_argumentation)
}
\arguments{
\item{explainer}{an iBreakDown explainer}

\item{display_argumentation}{displays argumentation mode}
}
\value{
a summary
}
\description{
Makes a summary for the description
}
