% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_break_down.R
\name{print.break_down}
\alias{print.break_down}
\title{Print Generic for Break Down Objects}
\usage{
\method{print}{break_down}(x, ..., digits = 3,
  rounding_function = round)
}
\arguments{
\item{x}{the model model of `break_down` class.}

\item{...}{other parameters.}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used.
See the \code{rounding_function} argument.}

\item{rounding_function}{function that is to used for rounding numbers.
It may be \code{\link{signif}} which keeps a specified number of significant digits.
Or the default \code{\link{round}} to have the same precision for all components.}
}
\value{
a data frame
}
\description{
Print Generic for Break Down Objects
}
\references{
Predictive Models: Visual Exploration, Explanation and Debugging \url{https://pbiecek.github.io/PM_VEE}
}
