% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_break_down_descriptions.R
\name{print.descriptions}
\alias{print.descriptions}
\title{Print Generic for Break Down Objects}
\usage{
\method{print}{descriptions}(x, ...)
}
\arguments{
\item{x}{description of `descriptions` class.}

\item{...}{other parameters.}
}
\value{
a string
}
\description{
Print Generic for Break Down Objects
}
