% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotD3_break_down.R
\name{plotD3}
\alias{plotD3}
\alias{plotD3.break_down}
\title{Plot Break Down Objects in D3 with r2d3 package.}
\usage{
plotD3(x, ...)

\method{plotD3}{break_down}(x, ..., baseline = NA, max_features = 10,
  min_max = NA, vcolors = DALEX::theme_drwhy_colors_break_down(),
  digits = 3, rounding_function = round)
}
\arguments{
\item{x}{the model model of `break_down`` class.}

\item{...}{other parameters.}

\item{baseline}{if numeric then veritical line will start in baseline.}

\item{max_features}{maximal number of features to be included in the plot. default value is 10.}

\item{min_max}{a range of OX axis. By deafult `NA` therefore will be extracted from the contributions of `x`. But can be set to some constants, usefull if these plots are used for comparisons.}

\item{vcolors}{named vector with colors.}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used.
See the \code{rounding_function} argument.}

\item{rounding_function}{function that is to used for rounding numbers.
It may be \code{\link{signif}} which keeps a specified number of significant digits.
Or the default \code{\link{round}} to have the same precision for all components.}
}
\value{
an `r2d3` object.
}
\description{
Plot Break Down Objects in D3 with r2d3 package.
}
\examples{
\dontrun{
## Not run:
# prepare dataset
library("titanic")
titanic <- titanic_train[,c("Survived", "Pclass", "Sex", "Age",
                            "SibSp", "Parch", "Fare", "Embarked")]
titanic$Survived <- factor(titanic$Survived)
titanic$Sex <- factor(titanic$Sex)
titanic$Embarked <- factor(titanic$Embarked)
titanic <- na.omit(titanic)

# prepare model
library("randomForest")
rf_model <- randomForest(Survived ~ .,  data = titanic)

# prepare explainer
library("DALEX")
rf_explain <- explain(rf_model, data = titanic,
                      y = titanic$Survived == "1", label = "RF")

# plor D3 explainers
library("iBreakDown")
rf_la <- local_attributions(rf_explain, titanic[2,])
rf_la
plotD3(rf_la)

rf_la <- local_attributions(rf_explain, titanic[3,])
rf_la
plotD3(rf_la, max_features = 4)
plotD3(rf_la, max_features = 4, min_max = c(0,1))
}
}
