% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_break_down.R
\name{print.break_down}
\alias{print.break_down}
\title{Print Generic for Break Down Objects}
\usage{
\method{print}{break_down}(x, ..., digits = 3, rounding_function = round)
}
\arguments{
\item{x}{an explanation created with \code{\link{break_down}}}

\item{...}{other parameters.}

\item{digits}{number of decimal places (round) or significant digits (signif) to be used.
See the \code{rounding_function} argument.}

\item{rounding_function}{a function to be used for rounding numbers.
This should be \code{\link{signif}} which keeps a specified number of significant digits or \code{\link{round}} (which is default) to have the same precision for all components.}
}
\value{
a data frame
}
\description{
Print Generic for Break Down Objects
}
\references{
Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://ema.drwhy.ai}
}
