% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_break_down_descriptions.R
\name{print.break_down_description}
\alias{print.break_down_description}
\title{Print Generic for Break Down Objects}
\usage{
\method{print}{break_down_description}(x, ...)
}
\arguments{
\item{x}{a description of \code{break_down_description} class.}

\item{...}{other parameters.}
}
\value{
a character
}
\description{
Print Generic for Break Down Objects
}
\references{
Explanatory Model Analysis. Explore, Explain and Examine Predictive Models. \url{https://pbiecek.github.io/ema}
}
